switch it everywhere in tests

+++ tests/pipelines/test_pipelines_mask_generation.py
@@ -83,7 +83,7 @@ class MaskGenerationPipelineTests(unittest.TestCase):
         image_processor=None,
         feature_extractor=None,
         processor=None,
-        torch_dtype="float32",
+        dtype="float32",
     ):
         image_segmenter = MaskGenerationPipeline(
             model=model,
@@ -91,7 +91,7 @@ class MaskGenerationPipelineTests(unittest.TestCase):
             feature_extractor=feature_extractor,
             image_processor=image_processor,
             processor=processor,
-            torch_dtype=torch_dtype,
+            dtype=dtype,
         )
         return image_segmenter, [
             "./tests/fixtures/tests_samples/COCO/000000039769.png",
